# openapi.yaml
openapi: 3.1.0
info:
  title: Pelevo API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /studio:
    get: { security: [{ bearerAuth: [] }], responses: { '200': { description: Verified creator profile }, '403': { description: Verified claim required } } }
...
  /studios:
    get: { security: [{ bearerAuth: [] }], responses: { '200': { description: Studios visible to the caller } } }
    post: { security: [{ bearerAuth: [] }], responses: { '201': { description: Creator studio created } } }
  /studios/{studio}:
    get: { security: [{ bearerAuth: [] }], responses: { '200': { description: Studio and members } } }
    put: { security: [{ bearerAuth: [] }], responses: { '200': { description: Owner updated studio } } }
    delete: { security: [{ bearerAuth: [] }], responses: { '200': { description: Owner deleted studio } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
...